feat: migrate block to css modules
