+++ packages/block/index.ts
@@ -1,2 +1,1 @@
-export { default as Block } from './Block'
-export * from './types'
+export * from './Block'
