+++ postcss.config.js
@@ -1,6 +1,9 @@
 /** @type {import('postcss-load-config').Config} */
 module.exports = {
   plugins: {
+    '@csstools/postcss-global-data': {
+      files: ['packages/theme/base/breakpoints.css'],
+    },
     'postcss-custom-media': {},
   },
 }
